Rework network configuration
network-config now generates a few different packages that can be used
for customizing what kind of connectivity you want out-of-the box in the
image. Based on this there is now three different images to use:

* core-image-qtipi-eth-client: DHCP over ethernet
* core-image-qtipi-wifi-ap: Pi act as AP, serving IPs on 192.168.9.x net
* core-image-qtipi-wifi-client: Pi connects to wifi

The SSID/Passkey for the two later options can be set in e.g.
local.conf:

QTIPI_SSID_pn-network-config = "myssid"
QTIPI_PASSKEY_pn-network-config = "mysecretpassword"